Drop the liveness sweep; cancellation is the remedy
The sweep existed because a worker can die mid-run, leaving an execution
RUNNING forever and -- via ra_agent_exec_one_active -- refusing every later
turn on that conversation. Cancellation fixes exactly that, and every
workflow that starts an execution now exposes it, so the sweep was the
belt to cancel's braces.

Deleting it removes the reason to guess a timeout, which is a question
with no good answer: a single provider call may retry inside the vendor SDK
for over an hour, and a tool handler may make several of those before it
returns. Three review rounds went into sizing that number, and every one
of them was a report of the sweep failing a healthy run.

Gone with it: the heartbeat contract the sweep read (the ambient
per-provider-call reporter, DatabaseAgentRecorder.heartbeat,
NestedRunHeartbeatRecorder, the proposal runner's boundary heartbeat, the
researcher-profile recorder parameter that only carried it), the beat
schedule, and the sweep's tests.

Kept: cancel, is_active, and the loop's check before each tool call, so a
cancelled turn still stops before it can edit a note.

The cost is that nothing cleans up after a dead worker unless a person
does, so execution rows can sit RUNNING indefinitely -- queries over
statuses and durations have to exclude them. In exchange the stuck state
is surfaced where someone can act on it: a busy conversation answers 409.
