+++ src/research_ai/services/agent_persistence/__init__.py
@@ -11,11 +11,8 @@ from .execution_service import (
     AgentExecutionService,
     AgentStaleRetryError,
 )
-from .liveness_service import (
-    AgentLivenessService,
-    ReclaimedExecutions,
-)
-from .recorder import DatabaseAgentRecorder, NestedRunHeartbeatRecorder
+from .liveness_service import AgentLivenessService
+from .recorder import DatabaseAgentRecorder
 from .retention_service import AgentRetentionService
 from .run_details_service import (
     AgentRunDetails,
@@ -34,8 +31,6 @@ __all__ = [
     "AgentRunDetailsService",
     "AgentStaleRetryError",
     "DatabaseAgentRecorder",
-    "NestedRunHeartbeatRecorder",
     "NoteAgentConversationService",
     "PreparedAgentExecution",
-    "ReclaimedExecutions",
 ]
